Updated constants.js (Some items from x21.0.0)
I added a bunch of items (not that many). These are the ones I have rn, + a bunch of ice shards and other things. I don't know how to add items that stack, without making 20 different lines and adding 20 pics of each stack, so I won't bother with that for now. I might add marks in the future. I have no idea if this is needed or not. I'm also getting up a renders.png for this.